segundo commit

+++ src/dao/productDao.js
@@ -1,27 +1,28 @@
-import { Schema, model } from 'mongoose'
-import { randomUUID } from 'crypto'
-import mongoosePaginate from 'mongoose-paginate-v2'
+import { Schema, model } from 'mongoose';
+import { randomUUID } from 'crypto';
+import mongoosePaginate from 'mongoose-paginate-v2';
 import { usersDao } from '../dao/index.js';
 
-const productSchema = new Schema ({
-    _id: { type: String, default: randomUUID},
-    title: { type: String, require: true },
-    description: { type: String, require: true },
-    code: { type: String, require: true },
-    price: { type: Number, require: true },
+const productSchema = new Schema({
+    _id: { type: String, default: randomUUID },
+    title: { type: String, required: true },
+    description: { type: String, required: true },
+    code: { type: String, required: true },
+    price: { type: Number, required: true },
     status: { type: Boolean, default: true },
-    stock: { type: Number, require: true },
-    category: { type: String, require: true },
+    stock: { type: Number, required: true },
+    category: { type: String, required: true },
     thumbnail: { type: String, default: '' },
     owner: { type: String, ref: 'users' } // Puedes cambiar el valor predeterminado según tus necesidades
 }, {
     strict: 'throw',
     versionKey: false,
-})
+});
 
-productSchema.plugin(mongoosePaginate)
+productSchema.plugin(mongoosePaginate);
 
-export const Producto = model('products', productSchema )
+export const Producto = model('products', productSchema);
+
 
 
 //------------------------------------------------------------------------------------------
@@ -77,16 +78,17 @@ export class ProductDao {
     };
 
     // Crear un nuevo producto
-    async crearProducto(usuarioId,nuevoProductoData){
+    async crearProducto(usuarioId, nuevoProductoData) {
         try {
-            // Asigna el propietario al ID del usuario premium
-            nuevoProductoData.owner = usuarioId;
-            const nuevoProducto = await Producto.create(nuevoProductoData);
-            return nuevoProducto;
+          // Asigna el propietario al ID del usuario premium
+          nuevoProductoData.owner = usuarioId;
+      
+          const nuevoProducto = await Producto.create(nuevoProductoData);
+          return nuevoProducto;
         } catch (error) {
-            throw new Error(`Error al crear un nuevo producto: ${error.message}`);
+          throw new Error(`Error al crear un nuevo producto: ${error.message}`);
         }
-    };
+    }
 
     // Actualizar un producto por ID
     async  actualizarProducto(productoId, usuarioId, newData) {
@@ -137,18 +139,41 @@ export class ProductDao {
       }
 
     // Eliminar un producto por ID
-    async  eliminarProducto(productoId){
+    async eliminarProducto(productoId, usuarioId) {
         try {
-            const delProducto = await Producto.findByIdAndDelete(productoId);
+            // Verificar si el usuario es admin o premium antes de permitir la eliminación
+            const usuario = await usersDao.findByIdUser(usuarioId);
+    
+            if (!usuario) {
+                throw new Error('Usuario no encontrado');
+            }
+    
+            // Verificar si el usuario tiene el rol de admin o premium
+            if (usuario.rol !== 'admin' && usuario.rol !== 'premium') {
+                throw new Error('No tienes permisos para eliminar productos');
+            }
+    
+            // Verificar si el producto existe
+            const producto = await Producto.findById(productoId);
 
+            console.log('el producto lo creo el siguiente usuario ', producto.owner)
+            console.log('el producto lo quiere eliminar el usuario ',usuarioId)
+    
+            // Verificar si el usuario tiene permisos para eliminar este producto
+            if (producto.owner.toString() !== usuarioId) {
+                throw new Error('No tienes permisos para modificar este producto');
+            }
+    
+            const delProducto = await Producto.findByIdAndDelete(productoId);
+    
             if (!delProducto) {
                 throw new Error(`El producto con id ${productoId} no se encontró`);
             }
-
+    
             return delProducto;
         } catch (error) {
             throw new Error(`Error al eliminar el producto por ID: ${error.message}`);
         }
-    };
+    }
 
 }
